cr_frontend/requirements.txt
pandas==1.3.0
requests
Flask==2.0.2
gunicorn==20.1.0
google-api-python-client
firebase-admin==5.2.0
structlog==21.5.0